+++ src/hotel/infrastructure/repository/hotel-repository.ts
@@ -1,8 +1,11 @@
-import { Injectable, Logger } from "@nestjs/common";
+import { Injectable, Logger, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { HotelRepositoryInterface } from "../../domain/repository/hotel-repository.interface";
 import { Hotel } from "../entity/hotel";
+import { HotelModel } from "src/hotel/domain/model/hotel-model";
+import { HotelMapper } from "../mapper/hotel-mapper";
+import { Constants } from "src/common/constants";
 
 @Injectable()
 export class HotelRepository implements HotelRepositoryInterface {
@@ -14,28 +17,42 @@ export class HotelRepository implements HotelRepositoryInterface {
         private readonly hotelRepository: Repository<Hotel>,
       ) {}
 
-    async get(): Promise<Hotel[]>{
-        return await this.hotelRepository.find();
+    async get(): Promise<HotelModel[]>{
+        const hotel = await this.hotelRepository.find();
+        return hotel.map(HotelMapper.entityToModel);
     }
 
-    async getById(id: string): Promise<Hotel | null>{
-        return await this.hotelRepository.findOneBy({id: id});
+    async getById(id: string): Promise<HotelModel>{
+        const hotel = await this.hotelRepository.findOneBy({id: id});
+        if(!hotel){
+            throw new NotFoundException(Constants.hotelNotFound);
+        }
+        return HotelMapper.entityToModel(hotel);
     }
 
-    async getByCountryAndCity(country: string, city: string): Promise<Hotel[]>{
-        return await this.hotelRepository.findBy({country: country, city: city});
+    async getByCountryAndCity(country: string, city: string): Promise<HotelModel[]>{
+        const hotel = await this.hotelRepository.findBy({country: country, city: city});
+        return hotel.map(HotelMapper.entityToModel);
     }
 
-    async create(hotel: Hotel): Promise<Hotel>{
-        this.hotelRepository.create(hotel);
-        return await this.hotelRepository.save(hotel);
+    async create(hotelModel: HotelModel): Promise<HotelModel>{
+        const hotel = HotelMapper.modelToEntity(hotelModel);
+        this.hotelRepository.create(hotelModel);
+        const response = await this.hotelRepository.save(hotel);
+        return HotelMapper.entityToModel(response);        
     }
 
-    async update(hotel: Hotel): Promise<Hotel>{
-        return await this.hotelRepository.save(hotel);
+    async update(hotelModel: HotelModel): Promise<HotelModel>{
+        const hotel = HotelMapper.modelToEntity(hotelModel);
+        const response = await this.hotelRepository.save(hotel);
+        return HotelMapper.entityToModel(response);
     }
 
-    async delete(hotel: Hotel): Promise<void>{
+    async delete(id: string): Promise<void>{
+        const hotel = await this.hotelRepository.findOneBy({id:id});
+        if(!hotel){
+            throw new NotFoundException(Constants.hotelNotFound);
+        }
         await this.hotelRepository.delete(hotel);
     }
 }
